+++ resources/js/Pages/AdicionarProduto.jsx
@@ -1,0 +1,12 @@
+import PrimaryButton from '@/Components/PrimaryButton';
+import AuthenticatedLayout from '../Layouts/AuthenticatedLayout';
+export default function TodoProduto()
+{
+  return (
+    <>
+    <AuthenticatedLayout>
+      <div>ola add</div>
+    </AuthenticatedLayout>
+    </>
+  )
+}