+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\AdicionarProdutoController;
 use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\TipoUsuarioController;
 use Illuminate\Foundation\Application;
@@ -17,6 +18,10 @@ Route::get('/', function () {
 Route::get('/dashboard', [TipoUsuarioController::class, 'tipoUsuario'])->middleware(['admin', 'verified'])->name('dashboard');
 
 Route::get("/todoproduto" , [TipoUsuarioController::class, "todoProduto"])->middleware(["admin" , 'verified'])->name('todoproduto');
+Route::get("/todoproduto/adicionarproduto", [AdicionarProdutoController::class , "AdicionarProduto"])->middleware(["admin" , "verified"])->name("adicionarProduto");
+
+
+
 
 Route::get('/minha-conta' , [TipoUsuarioController::class , 'minhaConta'])->middleware(['auth', 'verified'])->name('minha-conta');
 Route::middleware('auth')->group(function () {